+++ projects/frontend/es-app/src/guis/menus/MenuPads.h
@@ -48,6 +48,7 @@ class MenuPads : public Menu
       PadOSD,
       PadOSDType,
       AutoPairOnBoot,
+      Rumble,
     };
 
     //! Pad lists