+++ projects/manager3/src/stores/configuration/controllers.ts
@@ -70,6 +70,9 @@ export const useControllersStore = defineStore('controllers', {
       'bluetooth.autopaironboot': {
         value: true,
       },
+      rumble: {
+        value: false,
+      },
     },
   } as ControllersStoreState),