+++ RELEASE-NOTES.md
@@ -27,6 +27,7 @@ as release notes for end user on a Recalbox upgrade.
 - Add sort by playing time & sort by regions
 - Added options into configgen for Amiga systems
 - Force analog mode on Naomi, Naomi GD and Naomi 2
+- Add rumble option in frontend
 
 ### Fixes
 - Fix hidden and/or empty being shown after some user operations